Add devel-signed TMD hashes for Health & Safety.
This version uses the hashes from the official version of Health & Safety
from the dev CDN.

This allows installation on development units.

Autoboot isn't permitted due to system.xml's hash not matching.
Probably shouldn't set it on a development unit unless, at the very
least, the application type is changed to 0x90000001 (menu).